Add new csv and html output options.
The csv option is a simple wrapper over the old text one, but adds json escaping for values
The HTMl-table option just prints a simple <table> element, not a while page. It's got `data-status` attributes on the <tr> elements for custom styling, and tbody/tfoot elements as well. the tfoot has some summary data in it.