Add a new --summarize-only flag
This new flag controls the output that gets printed to the console, setting it will only print a summary instead of the full file list with all the annotation states listed.

You can set the flag on the cli with `--summary-only` or inside the package.json file with something like this:

```
{
  "flow-annotation-check": {
  	"summary_only": true
  }
}
```

This flag does not affect what gets saved directly into files (via the --html-file, --csv-file, or --junit-file flags). Those files will always have the full output which you can count & filter afterwards.