Create --level flag and aliases
Transform the previous --allow-weak flag into --level, so we can set
minimum level of strictness for files, and fail if it's not met. Also,
this ties into --list-files.

Include flowstrictlocal in visible status types, and split printStatusReport so we can use level

TODO: need to make list-files include weak files when `list-files=flow`
but `level=weak` as an upgrade warning.